ingest/requirements.txt
pandas
polars
pendulum
apache-airflow-providers-snowflake